增加经验条,存活时间和简单敌人

+++ ThePlagueForest/Assets/Scripts/GamePlay/CollisionHelper/CollisionHelper.cs
@@ -39,7 +39,7 @@ public static class CollisionHelper
             MyCollider collider2=enemy.GetCollider();
             if(IsColliding(collider1,collider2))
             {
-                int points=enemy.GetCurrentPropertySheet().GetAttack()/2;
+                int points=(int)(enemy.GetCurrentPropertySheet().GetAttack()/2);
                 DamageInfo damageInfo=new DamageInfo(enemy,player,points,null,null);
                 enemy.OnCollidePlayer();
                 FightSystem.Damage(damageInfo);
